Move district hierarchies from analysis -> merge_coded
