+++ survey_merge_coded/survey_merge_coded.py
@@ -2,7 +2,7 @@ import argparse
 import time
 from os import path
 
-from core_data_modules.cleaners import CharacterCleaner, Codes
+from core_data_modules.cleaners import CharacterCleaner, Codes, somali
 from core_data_modules.traced_data import Metadata
 from core_data_modules.traced_data.io import TracedDataJsonIO, TracedDataCodaIO, TracedDataTheInterfaceIO
 from core_data_modules.util import IOUtils
@@ -69,6 +69,32 @@ if __name__ == "__main__":
     for td in data:
         if td["district_coded"] == "other":
             td.append_data({"district_coded": "NC"}, Metadata(user, Metadata.get_call_location(), time.time()))
+
+    def convert_nc(value):
+        if value == Codes.NOT_CODED:
+            return "NC"
+        return value
+
+    # Set district/region/state/zone codes from the coded district field.
+    for td in data:
+        if not somali.DemographicCleaner.is_location_code(td["district_coded"]) and \
+                td["district_coded"] != Codes.STOP and \
+                td["district_coded"] != "NC" and td["district_coded"] is not None:
+            print("Unknown district: '{}'".format(td["district_coded"]))
+
+        td.append_data({
+            "district_coded": convert_nc(somali.DemographicCleaner.district_for_location_code(td["district_coded"])),
+            "region_coded": convert_nc(somali.DemographicCleaner.region_for_location_code(td["district_coded"])),
+            "state_coded": convert_nc(somali.DemographicCleaner.state_for_location_code(td["district_coded"])),
+            "zone_coded": convert_nc(somali.DemographicCleaner.zone_for_location_code(td["district_coded"])),
+            "district_coda": Codes.TRUE_MISSING if td["district_review"] == Codes.TRUE_MISSING else td["district_coded"]
+        }, Metadata(user, Metadata.get_call_location(), time.time()))
+
+        # If we failed to find a zone after searching location codes, try inferring from the operator code instead
+        if td["zone_coded"] == "NC":
+            td.append_data({
+                "zone_coded": convert_nc(somali.DemographicCleaner.zone_for_operator_code(td["operator"]))
+            }, Metadata(user, Metadata.get_call_location(), time.time()))
 
     # Merge manually coded activation Coda files into the cleaned dataset
     coda_file_path = path.join(coded_input_path, "esc4jmcna_activation_coded.csv")
